requirements.txt
appwrite==9.0.3
Scrapy==2.12.0
python-dotenv==1.1.0